pyproject.toml
[project]
name = "gerador-cvs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles==23.2.1",
    "fastapi==0.104.1",
    "httpx==0.25.2",
    "imapclient==2.3.1",
    "langchain==0.1.13",
    "langchain-community==0.0.29",
    "langchain-core==0.1.35",
    "langchain-openai==0.0.8",
    "loguru==0.7.2",
    "openai==1.12.0",
    "pandas==2.1.4",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
    "uvicorn[standard]==0.24.0",
]
